Esta version inserta los registros en Recepciones, actuliza las compras, presenta una exception al entrar a la lista de recepciones.

+++ MDSoft.Services/Dto/ComprasProductoDTO.cs
@@ -26,7 +26,7 @@ public partial class ComprasProductoDTO
 
     public string? RepNombre { get; set; }
 
-    public int? ComCantidadDetalle { get; set; }
+    public decimal? ComCantidadDetalle { get; set; }
 
     [Column("ComNCF")]
     [StringLength(50)]
