+++ MDSoft.Services/Repository/ComprasRepresentante.cs
@@ -8,7 +8,7 @@ namespace MDSoft.Tracking.Services.Repository
         public DateTime? ComFecha { get; set; }
         public string RepNombre { get; set; }
         public string ComReferencia { get; set; }
-        public int? ComCantidadDetalle { get; set; }
+        public decimal? ComCantidadDetalle { get; set; }
         public short? ComEstatus { get; set; }
     }
 }
